Improve tensor operations

+++ neurons/execution_layer/circuit.py
@@ -285,7 +285,7 @@ class CircuitEvaluationData:
         if response_times:
             log_circuit_metrics(response_times, verified_count, str(self.circuit))
 
-            response_tensor = torch.tensor(response_times)
+            response_tensor = torch.tensor(response_times, dtype=torch.float32)
             mean_response_time = torch.mean(response_tensor).item()
             max_response_time = torch.max(response_tensor).item()
             min_response_time = torch.min(response_tensor).item()
